step 6 changes
We now need to create a new zentral_mdm_artifact resource, link it to the MDM blueprint to put it in scope, and add a zentral_mdm_profile connected to it to load the mobileconfig.

Now that we have the artifact, we need to link it to a blueprint. In the mdm_default_blueprint.tf file,